backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.1.0
django-filter==23.3
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
python-dotenv==1.0.0